Threw a few things into endpoint classes
