+++ api/ExpressedRealms.Admin.API/AdminEndpoints/AdminEndpoints.cs
@@ -1,13 +1,13 @@
 using ExpressedRealms.Admin.API.AdminEndpoints.Dtos;
+using ExpressedRealms.Admin.API.AdminEndpoints.GetUsers;
 using ExpressedRealms.Admin.API.AdminEndpoints.Request;
 using ExpressedRealms.Admin.API.AdminEndpoints.Response;
-using ExpressedRealms.Admin.Repository;
+using ExpressedRealms.Admin.API.AdminEndpoints.UpdateUserRoles;
 using ExpressedRealms.Admin.Repository.ActivityLogs;
 using ExpressedRealms.Authentication;
 using ExpressedRealms.DB.UserProfile.PlayerDBModels.Roles;
 using ExpressedRealms.DB.UserProfile.PlayerDBModels.UserSetup;
 using ExpressedRealms.Server.Shared;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
@@ -28,82 +28,11 @@ public static class AdminEndpoints
             .WithOpenApi();
 
         endpointGroup
-            .MapGet(
-                "users",
-                [Authorize]
-                async Task<Ok<UserListResponse>> (IUsersRepository repository) =>
-                {
-                    var users = await repository.GetUsersAsync();
-
-                    return TypedResults.Ok(
-                        new UserListResponse()
-                        {
-                            Users = users
-                                .Select(x => new UserListItem()
-                                {
-                                    Id = x.Id,
-                                    Email = x.Email,
-                                    Username = x.Username,
-                                    Roles = x.Roles,
-                                    IsDisabled = x.IsDisabled,
-                                    EmailConfirmed = x.EmailConfirmed,
-                                    LockedOut = x.LockedOut,
-                                    LockedOutExpires = x.LockOutExpires,
-                                })
-                                .OrderBy(x => x.Email)
-                                .ToList(),
-                        }
-                    );
-                }
-            )
+            .MapGet("users", GetUsersEndpoint.Execute)
             .RequireAuthorization();
 
         endpointGroup
-            .MapPut(
-                "user/{userid}/role",
-                async Task<Results<NoContent, NotFound, BadRequest<string>>> (
-                    string userId,
-                    UpdateUserRoleRequest dto,
-                    UserManager<User> userManager,
-                    RoleManager<Role> roleManager,
-                    SignInManager<User> signInManager
-                ) =>
-                {
-                    var user = await userManager.FindByIdAsync(dto.UserId);
-
-                    if (user == null)
-                    {
-                        return TypedResults.NotFound();
-                    }
-
-                    // Ensure the role exists before assigning
-                    if (!await roleManager.RoleExistsAsync(dto.RoleName))
-                    {
-                        return TypedResults.NotFound();
-                    }
-
-                    if (dto.IsEnabled)
-                    {
-                        var result = await userManager.AddToRoleAsync(user, dto.RoleName);
-                        if (result.Succeeded)
-                        {
-                            return TypedResults.NoContent();
-                        }
-                    }
-                    else
-                    {
-                        var result = await userManager.RemoveFromRoleAsync(user, dto.RoleName);
-                        if (result.Succeeded)
-                        {
-                            return TypedResults.NoContent();
-                        }
-                    }
-
-                    await signInManager.RefreshSignInAsync(user);
-
-                    return TypedResults.BadRequest<string>("The role was not updated.");
-                }
-            )
+            .MapPut("user/{userid}/role", UpdateUserRoleEndpoint.Execute)
             .RequireAuthorization();
 
         endpointGroup
@@ -152,7 +81,7 @@ public static class AdminEndpoints
                     var roles = allRoles
                         .Select(x => new UserRoleDto()
                         {
-                            Name = x.Name,
+                            Name = x.Name!,
                             IsEnabled = userRoles.Any(y => y == x.Name),
                         })
                         .ToList();
