+++ api/ExpressedRealms.Admin.API/AdminEndpoints/GetUsers/UserListItem.cs
@@ -1,4 +1,4 @@
-namespace ExpressedRealms.Admin.API.AdminEndpoints.Dtos;
+namespace ExpressedRealms.Admin.API.AdminEndpoints.GetUsers;
 
 public class UserListItem
 {
