+++ api/ExpressedRealms.Admin.API/AdminEndpoints/GetUsers/UserListResponse.cs
@@ -1,0 +1,6 @@
+namespace ExpressedRealms.Admin.API.AdminEndpoints.GetUsers;
+
+public class UserListResponse
+{
+    public List<UserListItem> Users { get; set; } = new();
+}
