--- api/ExpressedRealms.Admin.API/AdminEndpoints/Response/UserListResponse.cs
@@ -1,8 +1,0 @@
-using ExpressedRealms.Admin.API.AdminEndpoints.Dtos;
-
-namespace ExpressedRealms.Admin.API.AdminEndpoints.Response;
-
-public class UserListResponse
-{
-    public List<UserListItem> Users { get; set; } = new();
-}
